server: correct i/o-streams
Just noticed that these can be generated by ClientConnect itself just
using connectionSocket. The main server does not need to do that as all
relevant information is contained within the socket.

Maybe we can change the constructor of ClientConnect to
	ClientConnect(DBQuery, ServerParser, Socket)

However this would require to rewrite tests. That could potentially be
painful.